refactor: 전략 패턴 도입
Before : Car 인터페이스를 구현한 RandomMovingCar로 움직임 정의로 일반화 과잉 문제 발생

After : 자동차의 움직임인 move()메서드만 전략패턴으로 추상화 향상
