+++ src/main/java/racingcar/controller/RacingController.java
@@ -5,7 +5,8 @@ import java.util.List;
 import racingcar.dto.RacingRequestDto;
 import racingcar.dto.RacingResponseDto;
 import racingcar.model.car.Car;
-import racingcar.model.car.RandomMovingCar;
+import racingcar.model.car.strategy.MoveBehavior;
+import racingcar.model.car.strategy.RandomMove;
 import racingcar.model.racing.Racing;
 import racingcar.model.result.RoundResult;
 import racingcar.util.CarNameParser;
@@ -37,10 +38,10 @@ public class RacingController {
     }
 
     private Racing initRacing(RacingRequestDto racingRequest) {
+        MoveBehavior behavior = new RandomMove();
         List<String> carNames = CarNameParser.parseCarName(racingRequest.rawCarNames());
         List<Car> cars = carNames.stream()
-                .map(RandomMovingCar::new)
-                .map(car -> (Car) car)
+                .map(carName -> new Car(behavior, carName))
                 .toList();
         int roundsToRace = RaceRoundParser.parseRaceRound(racingRequest.rawRoundsToRace());
         return Racing.from(cars, roundsToRace);
