+++ src/main/java/racingcar/model/car/strategy/MoveBehavior.java
@@ -1,0 +1,5 @@
+package racingcar.model.car.strategy;
+
+public interface MoveBehavior {
+    int getMoveAmount();
+}
