+++ src/main/java/racingcar/model/car/strategy/RandomMove.java
@@ -1,0 +1,18 @@
+package racingcar.model.car.strategy;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+public class RandomMove implements MoveBehavior {
+
+    private static final int MIN_RANDOM_NUMBER = 0;
+    private static final int MAX_RANDOM_NUMBER = 9;
+    private static final int MOVEMENT_THRESHOLD = 4;
+
+    @Override
+    public int getMoveAmount() {
+        if (Randoms.pickNumberInRange(MIN_RANDOM_NUMBER, MAX_RANDOM_NUMBER) >= MOVEMENT_THRESHOLD) {
+            return 1;
+        }
+        return 0;
+    }
+}
